refactor(frontend): clean code
