+++ frontend/src/pages/main/_utils/index.ts
@@ -1,1 +1,2 @@
 export * from './useReady';
+export * from './useEvent';
