+++ frontend/src/pages/main/channel/_components/channel-list/index.ts
@@ -1,0 +1,1 @@
+export * from './channel-list';
